feat(web,dash): implemented new textarea and input
