--- packages/ui/src/components/Input/V2/index.ts
@@ -1,2 +1,0 @@
-export { Input, type InputProps } from "@/components/Input/V2/Input";
-export { type InputType } from "@/components/Input/V2/types";
